requirements.txt
Django==1.5.1
MySQL-python==1.2.4b4
Pillow
distribute==0.6.27
south==0.8.1
sorl-thumbnail
shortuuid
#-e git://github.com/johnboxall/django-paypal.git
mandrill
django-mandrill
pytz
circus
circus-web
chaussette
django-sslify
xlsxwriter
